openai_backend/requirements.txt
Django==3.2
requests==2.26.0
psycopg2-binary==2.9.2
dj-database-url==0.5.0
environs
djangorestframework
openai
django-cors-headers